# backend/openapi.yaml
openapi: 3.0.3
info:
  title: VSmart Tech API
  version: 1.0.0
  description: Modular monolith CRM API. Public entry via Nginx; workers are not HTTP.
servers:
  - url: http://localhost:8081
    description: Docker Compose (Nginx)
paths:
...
  /api/v1/admin/tickets:
    get:
      summary: Admin list tickets
      tags: [tickets]
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Ticket list
  /api/v1/support/tickets:
    get:
      summary: Support engineer tickets
      tags: [tickets]
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Ticket list
...
  /api/v1/customer/tickets:
    get:
      summary: Customer tickets
      tags: [tickets]
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Ticket list
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...